requirements.txt
blosc2==2.0.0
certifi==2023.7.22
charset-normalizer==3.3.0
contourpy==1.1.1
cycler==0.12.0
Cython==3.0.2
filelock==3.12.4
fonttools==4.43.0
idna==3.4
importlib-resources==6.1.0
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
mplhep==0.3.29
mplhep-data==0.0.3
mpmath==1.3.0
msgpack==1.0.7
networkx==3.1
numexpr==2.8.7
numpy==1.26.0
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
psutil==5.9.5
py-cpuinfo==9.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
sympy==1.12
tables==3.8.0
termcolor==2.3.0
threadpoolctl==3.2.0
torch==2.0.1
torch-geometric==2.3.1
torch-scatter==2.1.1
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
uhi==0.3.3
urllib3==2.0.6
zipp==3.17.0